requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
appnope
asttokens
attrs
backcall
beautifulsoup4
bleach
cachetools
certifi
charset-normalizer
click
dataclasses-json  # 기존 0.6.7 버전 대신 호환되는 버전
decorator
defusedxml
docopt
executing
fastapi
fastjsonschema
filelock
frozenlist
fsspec
google-api-core
google-auth
google-cloud-vision
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httpx
huggingface-hub
idna
ipython
jedi
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
langchain
langsmith
MarkupSafe
marshmallow
matplotlib-inline
mistune
mpmath
multidict
mypy-extensions
nbclient
nbconvert
nbformat
networkx
numexpr
numpy
orjson
packaging
pandocfilters
parso
pexpect
pickleshare
pillow
pipreqs
platformdirs
prompt_toolkit
propcache
proto-plus
protobuf
ptyprocess
pure_eval
pyasn1
pyasn1_modules
pydantic
pydantic_core
Pygments
python-dateutil
python-dotenv
python-multipart
PyYAML
pyzmq
referencing
regex
requests
requests-toolbelt
rpds-py
rsa
safetensors
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
sympy
tenacity
tinycss2
tokenizers
torch
tornado
tqdm
traitlets
transformers
typing-inspect
typing_extensions
urllib3
uvicorn
wcwidth
webencodings
yarg
yarl